OpenAI API refactoring
